Added Created date in client create and update form

+++ client_bills/models.py
@@ -12,12 +12,12 @@ class ClientBill(models.Model):
     is_draft = models.BooleanField(default=False)
     bill_date = models.DateField()
     balance = models.IntegerField(default=0)
-    billed_amount = models.IntegerField(default=0)
+    billed_amount = models.FloatField(default=0)
     payment = models.ForeignKey(ClientPayment, blank=True, null=True, on_delete=models.CASCADE)
 
 class BillDetail(models.Model):
     bill = models.ForeignKey(ClientBill, on_delete=models.CASCADE)
     item = models.ForeignKey(Item, on_delete=models.CASCADE)
     unit = models.CharField(max_length=1, choices=[("k", "KG"), ("i", "Count")])
-    rate = models.IntegerField()
-    item_count = models.IntegerField()
+    rate = models.FloatField()
+    item_count = models.FloatField()
